requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.4
django-celery-beat==2.6.0
python-decouple==3.8
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
